[Feature] Add backup utility for user data
Implemented a backup utility that backs up all user data located in the 'groups' folder every day,
with a maximum of 5 backups stored. If the backup limit is exceeded, the oldest backup is removed.

In the future, consider moving static configuration options (such as backup frequency and max backups)
to the `config.yml` file, allowing for easier configuration.